change return value of `Channel::close` to `this`
There is no reason for `close` to return any particular value. The
intended result of calling `close` is purely its side-effect of closing
the channel.

Obliquely, returning `this` instead of `result` also fits better with
the upcoming “implicit receive” feature, as this would allow for the
expression `value = yield ch.close(result)`. Previously this was
malformed (yielding an arbitrary `result` had no defined meaning),
whereas now this evaluates to yielding a **receive** operation on the
closed channel, causing `result` to be assigned to `value` (via
`Process/run`) — which was the apparent goal of having `close` directly
return `result` in the first place.